leds: refactor uart interface
