+++ components/leds/interface.h
@@ -8,11 +8,9 @@
 #if CONFIG_LEDS_SPI_ENABLED
 # include "interfaces/spi.h"
 #endif
-/* XXX
 #if CONFIG_LEDS_UART_ENABLED
 # include "interfaces/uart.h"
 #endif
-*/
 
 union leds_interface_state {
 #if CONFIG_LEDS_SPI_ENABLED
@@ -20,5 +18,8 @@ union leds_interface_state {
 #endif
 #if CONFIG_LEDS_I2S_ENABLED
   struct leds_interface_i2s i2s;
+#endif
+#if CONFIG_LEDS_UART_ENABLED
+  struct leds_interface_uart uart;
 #endif
 };
